Update .travis.yml to resolve docker failure
Travis builds are failing because the use of the `-p` flag in the `docker login` command causes a warning that requires user input, making the build time out. This switches the `docker login` command to match [what Travis recommends](https://docs.travis-ci.com/user/docker/#Pushing-a-Docker-Image-to-a-Registry) and should resolve the issue.